mentions
added full(?) functionality for mentions. If no channel mentions are made, it defaults to the entire server. if no user mentions are made, it defaults to the user of the command.

Basic functionality is now accomplished. Further functionality would consist of a persistent shelve file, and graphical representations over time.